+++ apps/ui/src/components/sidebar/views/info/how-it-works.tsx
@@ -1,4 +1,4 @@
-import { LinkedText } from "../utils/linked-text";
+import { LinkedText } from "../../../utils/linked-text";
 
 export function HowItWorks() {
   return (
